chore(agent): [Step 3/9] Harden desktop OAuth service and platform split
PLAN_REF: 640662d320886107154dee6729268ee3912823c5
PREVIOUS_STEP: 0981549a — Enforced secure profile-scoped OAuth token storage

## Completed in This Step
- Split OAuth service behind a conditional export so shared imports are web-safe and the desktop implementation owns dart:io usage.
- Added a non-IO OAuth stub that fails safely on unsupported platforms while still allowing cached credential cleanup.
- Changed loopback redirects to an exact /oauth/callback path on a held port-0 listener, removing the free-port TOCTOU pattern.
- Rejected mismatched OAuth state instead of accepting authorization codes.
- Added first-terminal callback handling and closes the loopback server immediately after a terminal callback.
- Closed HttpClient instances in metadata, DCR, redirect discovery, and token exchange paths.
- Tightened Cloudflare Access host checks and removed sensitive callback URL/state/code logging.
- Updated DCR metadata to use the CodeWalk upstream client URI and include the protected resource.

## References Used
- final_plan.md Phase 3 and Phase 4 callback/platform split requirements.
- Cloudflare Managed OAuth contract proof: PKCE S256, loopback redirect, DCR, refresh token semantics.
- ADR-023: Cloudflare OAuth remains reverse-proxy access-layer auth.

## Debts and Warnings
- Dio/SSE auth ownership and AppProvider cached credential activation remain for Step 4.
- UX/i18n platform gating remains for Step 5.
- Targeted callback tests still need to be added in Step 6.

## Progress (ref: 640662d320886107154dee6729268ee3912823c5)
- [x] Step 1: Update PR branch against origin/main and resolve conflicts while preserving contributor commits.
- [x] Step 2: Fix secure OAuth credential persistence and profile-scoped credential model.
- [x] Step 3: Harden OAuth service callback, discovery, DCR, refresh, platform split, and logging.
- [ ] Step 4: Refactor Dio/SSE auth ownership and AppProvider profile lifecycle.
- [ ] Step 5: Gate and localize UX for Cloudflare OAuth.
- [ ] Step 6: Add targeted tests for final_plan.md failure modes.
- [ ] Step 7: Run make check and required scans, fix failures.
- [ ] Step 8: Run reviewer1/reviewer2 on all code commits and apply judge-approved fixes in a loop.
- [ ] Step 9: Update docs through subagents, push to PR #37, monitor CI, and merge to main.

## Next Step
chore(agent): [Step 4/9] Refactor Dio/SSE auth ownership and AppProvider profile lifecycle
